message_waiter/requirements.txt
python-telegram-bot==20.0
httpx==0.23.1
pydantic==1.9.0